[Flyway][Eason] feature: implement should_throw_no_data_exception_when_delete_employee_by_id_given_id test

+++ src/main/java/com/thoughtworks/springbootemployee/service/EmployeeService.java
@@ -58,9 +58,11 @@ public class EmployeeService {
         }
     }
 
-    public boolean deleteById(int id) {
+    public boolean deleteById(int id) throws NoSuchDataException {
+        if (!employeeRepository.findById(id).isPresent()) {
+            throw new NoSuchDataException();
+        }
         employeeRepository.deleteById(id);
-
-        return !employeeRepository.findById(id).isPresent();
+        return true;
     }
 }
